+++ src/main/java/tekcays_addon/api/utils/TKCYAValues.java
@@ -61,9 +61,9 @@ public class TKCYAValues {
     public static final float FLUID_MULTIPLIER_PRESSURE = 100.0f;
     public static final String NO_FLUID = "";
     public static final int ABSOLUTE_VACUUM = 0;
-    public static final long ATMOSPHERIC_PRESSURE = 101300;
+    public static final int ATMOSPHERIC_PRESSURE = 101300;
     public static final float OXYGEN_IN_AIR_RATIO = 0.21f;
-    public static final long MAX_PRESSURE = ATMOSPHERIC_PRESSURE * 2000;
+    public static final int MAX_PRESSURE = ATMOSPHERIC_PRESSURE * 2000;
     public static final float EU_TO_PU = 0.40F;
     public static final float EU_TO_VU = 0.30F;
     /**
