+++ src/main/java/tekcays_addon/common/metatileentities/single/MetaTileEntityElectricPressureCompressor.java
@@ -19,7 +19,7 @@ public class MetaTileEntityElectricPressureCompressor extends ElectricPressureCo
     private IPressureContainer pressureContainer;
     private int fluidCapacity;
     private int tierMultiplier = (getTier() * getTier() + 1);
-    private long pressure;
+    private int pressure;
 
     public MetaTileEntityElectricPressureCompressor(ResourceLocation metaTileEntityId, int tier) {
         super(metaTileEntityId, false, tier);
@@ -96,7 +96,7 @@ public class MetaTileEntityElectricPressureCompressor extends ElectricPressureCo
     }
 
     @Override
-    public long getPressure() {
+    public int getPressure() {
         return pressureContainer.getPressure();
     }
